+++ custom_tools/geometry_tools.py
@@ -1,6 +1,5 @@
 import arcpy
-from env_setup import environment_setup
-from custom_tools import custom_arcpy
+from custom_tools.general_tools import custom_arcpy
 from file_manager.n100.file_manager_rivers import River_N100
