+++ custom_tools/partition_iterator_state_based.py
@@ -8,7 +8,7 @@ from typing import Dict, Tuple, Literal
 
 import env_setup.global_config
 from env_setup import environment_setup
-from custom_tools import custom_arcpy
+from custom_tools.general_tools import custom_arcpy
 from custom_tools.decorators.timing_decorator import timing_decorator
 
 from input_data import input_n50
