+++ generalization/n100/building/calculate_point_values.py
@@ -1,5 +1,5 @@
 # Importing custom files relative to the root path
-from custom_tools import custom_arcpy
+from custom_tools.general_tools import custom_arcpy
 from env_setup import environment_setup
 from file_manager.n100.file_manager_buildings import Building_N100
 from constants.n100_constants import N100_SQLResources
