+++ generalization/n100/building/finalizing_buildings.py
@@ -2,7 +2,7 @@
 import arcpy
 
 # Importing custom files
-from custom_tools import custom_arcpy
+from custom_tools.general_tools import custom_arcpy
 from env_setup import environment_setup
 from file_manager.n100.file_manager_buildings import Building_N100
