+++ generalization/n100/building/not_in_use_models/iteration.py
@@ -2,7 +2,7 @@ import arcpy
 import os
 
 from env_setup import environment_setup
-from custom_tools import custom_arcpy
+from custom_tools.general_tools import custom_arcpy
 from file_manager.n100.file_manager_buildings import Building_N100
