+++ generalization/n100/building/not_in_use_models/old_resolve_building_conflicts.py
@@ -1,16 +1,12 @@
 import arcpy
-import os
 import time
 
 # Importing custom files relative to the root path
 import config
-from custom_tools import custom_arcpy
+from custom_tools.general_tools import custom_arcpy
 from env_setup import environment_setup
-from input_data import input_n50
-from input_data import input_n100
 from input_data.input_symbology import SymbologyN100
 from file_manager.n100.file_manager_buildings import Building_N100
-from env_setup import setup_directory_structure
 
 # Start timing
 start_time = time.time()
