+++ generalization/n100/building/point_displacement_with_buffer.py
@@ -9,7 +9,7 @@ from constants.n100_constants import N100_Symbology, N100_SQLResources, N100_Val
 from custom_tools.polygon_processor import PolygonProcessor
 from env_setup import environment_setup
 from custom_tools.decorators.timing_decorator import timing_decorator
-from custom_tools import custom_arcpy
+from custom_tools.general_tools import custom_arcpy
 
 
 @timing_decorator
