+++ generalization/n100/building/removing_overlapping_points.py
@@ -3,7 +3,7 @@
 # Importing custom files
 
 # Import custom modules
-from custom_tools import custom_arcpy
+from custom_tools.general_tools import custom_arcpy
 from env_setup import environment_setup
 from custom_tools.compare_feature_classes import compare_feature_classes
 from custom_tools.polygon_processor import PolygonProcessor
