+++ generalization/n100/building/simplify_polygons.py
@@ -5,7 +5,7 @@ import arcpy
 from env_setup import environment_setup
 from input_data import input_n50
 from file_manager.n100.file_manager_buildings import Building_N100
-from custom_tools import custom_arcpy
+from custom_tools.general_tools import custom_arcpy
 from constants.n100_constants import N100_Values
 
 # Importing timing decorator
