+++ generalization/n100/river/not_in_use_models/centerline_pruning.py
@@ -1,11 +1,9 @@
 import arcpy
 import networkx as nx
 import os
-from itertools import combinations
 import math
 
 from env_setup import environment_setup
-from custom_tools import custom_arcpy
 from file_manager.n100.file_manager_rivers import River_N100
